backend/requirements.txt
flask
flask-sqlalchemy
flask-login
flask-cors
werkzeug
Pillow
